scrapping and downloading papers
This code can download and scrap the papers. However, please note that the papers need to be open access so not any link will work. My errors were mainly because I was using open-access PDFs, but the website doesn't authorize scrapping, and some can be illegal. 
Please try it and let me know if you found any errors or if I missed anything. 
I am a bit confused because, for some steps to work, I have to do the step before it to make sure that mine will work. So please let me know if I missed something.